pyproject.toml
[project]
name = "paperless-orchestrator"
version = "0.1.0"
description = "An AI-powered document orchestrator for Paperless-NGX with a Streamlit UI."
authors = ["Gemini"]
requires-python = ">=3.9"
dependencies = [
    "google-adk>=1.0.0",
    "google-generativeai",
    "httpx",
    "python-dotenv",
    "pdfplumber",
    "streamlit>=1.50.0",
    "nest-asyncio",
    # Dependencies from the old agent project that might be needed
    "boto3",
    "google-cloud-storage",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99
